Upgrade husky to latest version, fix package.json
