+++ packages/sdk-dev/src/playground/index.html
@@ -6,5 +6,6 @@
   </head>
   <body>
     <div id="root"></div>
+    <script src="./index.tsx"></script>
   </body>
 </html>
